% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_campsis_suite_packages}
\alias{get_campsis_suite_packages}
\title{Get all packages from the Campsis suite.}
\usage{
get_campsis_suite_packages(include_engines = TRUE)
}
\arguments{
\item{include_engines}{include simulation engines (mrgsolve and rxode2), default is TRUE}
}
\value{
a character vector with the packages
}
\description{
Get all packages from the Campsis suite.
}
