% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_private_packages}
\alias{get_private_packages}
\title{Get the private packages from the Campsis suite.}
\usage{
get_private_packages()
}
\value{
a character vector of the private packages
}
\description{
Get the private packages from the Campsis suite.
}
