% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_public_packages}
\alias{get_public_packages}
\title{Get the public packages from the Campsis suite.}
\usage{
get_public_packages()
}
\value{
a character vector of the public packages
}
\description{
Get the public packages from the Campsis suite.
}
