% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getAvailableVersions}
\alias{getAvailableVersions}
\title{Get all available versions.}
\usage{
getAvailableVersions(as_date = FALSE)
}
\arguments{
\item{as_date}{return versions as dates, default is FALSE}
}
\description{
Get all available versions.
}
