% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getPrivatePackages}
\alias{getPrivatePackages}
\title{Get the private packages from the Campsis suite.}
\usage{
getPrivatePackages()
}
\value{
a character vector of the private packages
}
\description{
Get the private packages from the Campsis suite.
}
