% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getPublicPackages}
\alias{getPublicPackages}
\title{Get the public packages from the Campsis suite.}
\usage{
getPublicPackages()
}
\value{
a character vector of the public packages
}
\description{
Get the public packages from the Campsis suite.
}
