% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{loadEnv}
\alias{loadEnv}
\title{Load environment.}
\usage{
loadEnv(dir = getEnvDefaultDir(), version = getPackageVersion())
}
\arguments{
\item{dir}{temporary directory}

\item{version}{campsisverse version}
}
\description{
Load environment.
}
