% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{qualify}
\alias{qualify}
\title{Use environment.}
\usage{
qualify(packages, fullname, output_dir = ".")
}
\arguments{
\item{packages}{campsisverse version}

\item{...}{extra arguments}
}
\description{
Use environment.
}
