% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{renv_lock_0.0.1}
\alias{renv_lock_0.0.1}
\title{Renv lock file version '0.0.1'.}
\format{
An object of class \code{character} of length 1.
}
\usage{
renv_lock_0.0.1
}
\description{
Renv lock file version '0.0.1'.
}
\keyword{datasets}
