% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{renv_lock_1.0.0}
\alias{renv_lock_1.0.0}
\title{Renv lock file version '1.0.0'.}
\format{
An object of class \code{character} of length 1.
}
\usage{
renv_lock_1.0.0
}
\description{
Renv lock file version '1.0.0'.
}
\keyword{datasets}
