% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{renv_lock_241201}
\alias{renv_lock_241201}
\title{Renv lock file made on 241201.}
\format{
An object of class \code{character} of length 1.
}
\usage{
renv_lock_241201
}
\description{
Renv lock file made on 241201.
}
\keyword{datasets}
