% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{renv_lock_250209}
\alias{renv_lock_250209}
\title{Renv lock file made on 250209.}
\format{
An object of class \code{character} of length 1.
}
\usage{
renv_lock_250209
}
\description{
Renv lock file made on 250209.
}
\keyword{datasets}
