% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{restore}
\alias{restore}
\title{Restore environment.}
\usage{
restore(
  version = get_package_version(),
  all = FALSE,
  no_deps = FALSE,
  library = .libPaths(),
  ...
)
}
\arguments{
\item{version}{campsisverse version}

\item{all}{all packages, included private ones, default is FALSE. Reserved for Calvagone members only.}

\item{no_deps}{do not restore Campsis suite dependencies as specified in the lock file, default is FALSE.}

\item{library}{library paths to be used during restore, default is \code{.libPaths()}}

\item{...}{extra arguments}
}
\description{
Restore environment.
}
