% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{runQualification}
\alias{runQualification}
\title{Qualify the Campsis environment. Note that this is just a call to the campsisqual::runQualification function.}
\usage{
runQualification(packages, fullname, output_dir = getwd())
}
\arguments{
\item{packages}{campsisverse version}

\item{fullname}{full user name (firstname lastname)}

\item{output_dir}{output directory of the qualification report}
}
\description{
Qualify the Campsis environment. Note that this is just a call to the campsisqual::runQualification function.
}
