% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{uninstall}
\alias{uninstall}
\title{Uninstall the Campsis suite.}
\usage{
uninstall(all = FALSE)
}
\arguments{
\item{all}{all packages included private ones, default is FALSE}
}
\description{
Uninstall the Campsis suite.
}
