% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{use}
\alias{use}
\title{Use environment.}
\usage{
use(version = get_package_version(), all = FALSE, no_deps = FALSE, ...)
}
\arguments{
\item{version}{campsisverse version}

\item{all}{all packages, included private ones, default is FALSE. Reserved for Calvagone members only.}

\item{no_deps}{do not use Campsis suite dependencies as specified in the lock file, default is FALSE.}

\item{...}{extra arguments}
}
\description{
Use environment.
}
