% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getLockFile}
\alias{getLockFile}
\title{Get lock file.}
\usage{
getLockFile(version = getPackageVersion(), all = FALSE)
}
\arguments{
\item{version}{campsisverse version}

\item{all}{all packages, included private ones (authentication key needed), default is FALSE}
}
\description{
Get lock file.
}
