% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getLockFile}
\alias{getLockFile}
\title{Get lock file.}
\usage{
getLockFile(version = getPackageVersion())
}
\arguments{
\item{version}{campsisverse version}

\item{dir}{temporary directory}
}
\description{
Get lock file.
}
