% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getLockFile}
\alias{getLockFile}
\title{Get lock file.}
\usage{
getLockFile(
  version = getPackageVersion(),
  all = FALSE,
  no_deps = FALSE,
  prompt = TRUE,
  discard_renv = FALSE
)
}
\arguments{
\item{version}{campsisverse version}

\item{all}{all packages, included private ones (authentication key needed), default is FALSE}

\item{no_deps}{discard Campsis suite dependencies specified in the lock file, default is FALSE.}

\item{prompt}{prompt the user for input, default is TRUE (e.g. if campsisverse must be updated)}

\item{discard_renv}{discard renv package from the lock file, default is FALSE}
}
\description{
Get lock file.
}
