% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{restore}
\alias{restore}
\title{Restore environment.}
\usage{
restore(version = getPackageVersion(), all = FALSE, ...)
}
\arguments{
\item{version}{campsisverse version}

\item{all}{all packages, included private ones, default is FALSE. Reserved for Calvagone members only.}

\item{...}{extra arguments}
}
\description{
Restore environment.
}
