% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{restore}
\alias{restore}
\title{Restore environment.}
\usage{
restore(version = getPackageVersion(), ...)
}
\arguments{
\item{version}{campsisverse version}

\item{...}{extra arguments}
}
\description{
Restore environment.
}
