% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{restoreEnv}
\alias{restoreEnv}
\title{Restore environment.}
\usage{
restoreEnv(dir = dirname(tempdir()), version = "0.0.1")
}
\arguments{
\item{dir}{temporary directory}

\item{version}{campsisverse version}
}
\description{
Restore environment.
}
