% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{use}
\alias{use}
\title{Use environment.}
\usage{
use(version = getPackageVersion(), all = FALSE, ...)
}
\arguments{
\item{version}{campsisverse version}

\item{all}{all packages, included private ones (authentication key needed), default is FALSE}

\item{...}{extra arguments}
}
\description{
Use environment.
}
